switch from removing to disabling rsyslog
Because in principle (and practice), packages could depend on rsyslog.

Some Debian packages indirectly depend on it and cannot be satisfied by
systemd, but they only actually need a working /dev/log.  The solution
is proposed here:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=799549#57

In the meantime, let's avoid silently removing packages which the admin had
deliberately installed.